Fix userforms to 4.3.1
Userforms 4.3.x up until 4.3.5 requires silverstripe-australia/gridfieldextensions, but
after 4.3.5 it requires symbiote/silverstripe-gridfieldextensions. This conflicts with the
requirements in the recipe, so I'm fixing the userforms constraint to a stable tag. If we
need to release a new userforms version in future we can bump the tag in the next 1.6
recipe release.